Backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9
gunicorn>=21.2.0
python-decouple>=3.8 
whitenoise>=6.6.0 
python-dotenv